Revised thresholders, Flusser moments plugins

+++ imagingbook-plugins-book/src/main/java/Ch08_Binary_Regions/Major_Axis_Demo.java
@@ -3,7 +3,7 @@
  * image processing published by Springer-Verlag in various languages and editions.
  * Permission to use and distribute this software is granted under the BSD 2-Clause
  * "Simplified" License (see http://opensource.org/licenses/BSD-2-Clause).
- * Copyright (c) 2006-2022 Wilhelm Burger, Mark J. Burge. All rights reserved.
+ * Copyright (c) 2006-2023 Wilhelm Burger, Mark J. Burge. All rights reserved.
  * Visit https://imagingbook.com for additional details.
  ******************************************************************************/
 package Ch08_Binary_Regions;
@@ -104,8 +104,8 @@ public class Major_Axis_Demo implements PlugInFilter {	// TODO: convert to overl
 				Pnt2d xc = r.getCenter();
 				double x0 = xc.getX();
 				double y0 = xc.getY();
-				double x1 = x0 + DrawingScale * r.getProperty("dx");
-				double y1 = y0 + DrawingScale * r.getProperty("dy");
+				double x1 = x0 + DrawingScale * (double) r.getProperty("dx");
+				double y1 = y0 + DrawingScale * (double) r.getProperty("dy");
 				ola.addShape(xc.getShape(0.05 * DrawingScale));
 				ola.addShape(new Line2D.Double(x0, y0, x1, y1));
 			}
